@@ -3,52 +3,52 @@
     "type": "project",
     "license": "MIT",
     "require": {
-        "php": "^7.3",
+        "php": "^8.0",
         "ext-ctype": "*",
         "ext-iconv": "*",
         "doctrine/annotations": "^1.0",
-        "doctrine/doctrine-bundle": "^1.6.10|^2.0",
-        "doctrine/doctrine-migrations-bundle": "^1.3|^2.0",
-        "doctrine/orm": "^2.5.11",
-        "phpdocumentor/reflection-docblock": "^3.0|^4.0",
-        "sensio/framework-extra-bundle": "^5.1",
-        "symfony/asset": "5.2.*",
-        "symfony/console": "5.2.*",
-        "symfony/dotenv": "5.2.*",
-        "symfony/expression-language": "5.2.*",
+        "doctrine/doctrine-bundle": "^2.0",
+        "doctrine/orm": "^2.11",
+        "phpdocumentor/reflection-docblock": "^5.3",
+        "sensio/framework-extra-bundle": "^6.0",
+        "symfony/asset": "6.0.*",
+        "symfony/console": "6.0.*",
+        "symfony/dotenv": "6.0.*",
+        "symfony/expression-language": "6.0.*",
         "symfony/flex": "^1.1",
-        "symfony/form": "5.2.*",
-        "symfony/framework-bundle": "5.2.*",
+        "symfony/form": "6.0.*",
+        "symfony/framework-bundle": "6.0.*",
         "symfony/monolog-bundle": "^3.1",
-        "symfony/process": "5.2.*",
-        "symfony/property-access": "5.2.*",
-        "symfony/property-info": "5.2.*",
-        "symfony/security-bundle": "5.2.*",
-        "symfony/serializer": "5.2.*",
-        "symfony/twig-bundle": "5.2.*",
-        "symfony/validator": "5.2.*",
-        "symfony/web-link": "5.2.*",
-        "symfony/workflow": "5.2.*",
-        "symfony/yaml": "5.2.*"
+        "symfony/process": "6.0.*",
+        "symfony/property-access": "6.0.*",
+        "symfony/property-info": "6.0.*",
+        "symfony/runtime": "6.0.*",
+        "symfony/security-bundle": "6.0.*",
+        "symfony/serializer": "6.0.*",
+        "symfony/twig-bundle": "6.0.*",
+        "symfony/validator": "6.0.*",
+        "symfony/web-link": "6.0.*",
+        "symfony/workflow": "6.0.*",
+        "symfony/yaml": "6.0.*"
     },
     "require-dev": {
-        "easycorp/easy-log-handler": "^1.0.7",
-        "symfony/browser-kit": "5.2.*",
-        "symfony/css-selector": "5.2.*",
-        "symfony/debug-bundle": "5.2.*",
+        "symfony/browser-kit": "6.0.*",
+        "symfony/css-selector": "6.0.*",
+        "symfony/debug-bundle": "6.0.*",
         "symfony/maker-bundle": "^1.0",
-        "symfony/phpunit-bridge": "5.2.*",
-        "symfony/stopwatch": "5.2.*",
-        "symfony/var-dumper": "5.2.*",
-        "symfony/web-profiler-bundle": "5.2.*"
+        "symfony/phpunit-bridge": "6.0.*",
+        "symfony/stopwatch": "6.0.*",
+        "symfony/var-dumper": "6.0.*",
+        "symfony/web-profiler-bundle": "6.0.*"
     },
     "config": {
         "preferred-install": {
             "*": "dist"
         },
         "sort-packages": true,
-        "platform": {
-            "php": "7.3.17"
+        "allow-plugins": {
+            "symfony/flex": true,
+            "symfony/runtime": true
         }
     },
     "autoload": {
@@ -65,6 +65,7 @@
         "paragonie/random_compat": "2.*",
         "symfony/polyfill-ctype": "*",
         "symfony/polyfill-iconv": "*",
+        "symfony/polyfill-php80": "*",
         "symfony/polyfill-php74": "*",
         "symfony/polyfill-php73": "*",
         "symfony/polyfill-php72": "*",
@@ -92,7 +93,7 @@
     "extra": {
         "symfony": {
             "allow-contrib": false,
-            "require": "5.2.*"
+            "require": "6.0.*"
         }
     }
 }
