@@ -13,6 +13,6 @@ class HomepageControllerTest extends WebTestCase
         $crawler = $client->request('GET', '/');
 
         $this->assertEquals(200, $client->getResponse()->getStatusCode());
-        $this->assertContains('symfony/workflow', $crawler->filter('.container h1')->text());
+        $this->assertStringContainsString('symfony/workflow', $crawler->filter('.container h1')->text());
     }
 }
